moved script
