+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "@baltimorecounty/dotgov-components/lib/styles/dotgov.min.css";
-import Homepage from "./components/HomePage";
+import Homepage from "./pages/HomePage";
 
 function App() {
   return <Homepage />;
